Alteracao do tipo dos dados

+++ db/migrate/20150724001647_add_periodo_valor_parcelas_to_tabelaplano.rb
@@ -1,7 +1,7 @@
 class AddPeriodoValorParcelasToTabelaplano < ActiveRecord::Migration
   def change
     add_column :tabelaplanos, :periodo, :string
-    add_column :tabelaplanos, :valor, :double
+    add_column :tabelaplanos, :valor, :float
     add_column :tabelaplanos, :parcelas, :integer
   end
 end
